LayoutTests/http/tests/security/text-track-crossorigin-expected.txt

CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.
Tests loading cross-domain <track>.


Loading without Access-Control-Allow-Origin header, no "crossorigin" attribute on <video>
EVENT(load)
PASS: shouldLoad should be 'true' and is.
PASS: event.target should be '[object HTMLTrackElement]' and is.
PASS: trackElement.readyState should be '2' and is.


Loading without Access-Control-Allow-Origin header, setting video.crossorigin to "anonymous"
EVENT(error)
PASS: shouldLoad should be 'false' and is.
PASS: event.target should be '[object HTMLTrackElement]' and is.
PASS: trackElement.readyState should be '3' and is.


Loading with Access-Control-Allow-Origin header, leaving video.crossorigin as "anonymous"
EVENT(load)
PASS: shouldLoad should be 'true' and is.
PASS: event.target should be '[object HTMLTrackElement]' and is.
PASS: trackElement.readyState should be '2' and is.


Loading with Access-Control-Allow-Origin and Access-Control-Allow-Credentials headers, setting video.crossorigin to "use-credentials"
EVENT(load)
PASS: shouldLoad should be 'true' and is.
PASS: event.target should be '[object HTMLTrackElement]' and is.
PASS: trackElement.readyState should be '2' and is.


END OF TEST
